extract task functions
